refactor: more vitest
