+++ jest.config.js
@@ -21,7 +21,7 @@ module.exports = {
   ],
   modulePathIgnorePatterns: ['<rootDir>/package.json'],
   resetMocks: false,
-  setupFiles: ['./jest.setup.js', 'jest-localstorage-mock', './match-media-mock.js'],
+  setupFiles: ['./jest.setup.js', 'jest-localstorage-mock'],
   setupFilesAfterEnv: ['@testing-library/jest-dom/extend-expect'],
   transform: {
     '^.+\\.tsx?$': ['ts-jest', { tsconfig: 'tsconfig.json' }],
