+++ package.json
@@ -69,7 +69,6 @@
     "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.4.1",
     "jest-localstorage-mock": "^2.4.18",
-    "jest-websocket-mock": "^2.1.0",
     "jsdom": "^26.1.0",
     "mockjs": "^1.1.0",
     "react": "^19.1.0",
@@ -83,6 +82,7 @@
     "ts-jest": "^29.1.1",
     "typescript": "^5.8.3",
     "vitest": "^3.2.4",
+    "vitest-websocket-mock": "^0.5.0",
     "webpack": "^5.99.9",
     "webpack-cli": "^6.0.1",
     "webpack-merge": "^6.0.1"
