--- packages/hooks/src/useToggle/__tests__/index.test.ts
@@ -1,42 +1,0 @@
-import { renderHook, act } from '@testing-library/react';
-import useToggle from '../index';
-
-const callToggle = (hook: any) => {
-  act(() => {
-    hook.result.current[1].toggle();
-  });
-};
-
-describe('useToggle', () => {
-  it('test on init', async () => {
-    const hook = renderHook(() => useToggle());
-    expect(hook.result.current[0]).toBeFalsy();
-  });
-
-  it('test on methods', async () => {
-    const hook = renderHook(() => useToggle('Hello'));
-    expect(hook.result.current[0]).toBe('Hello');
-    callToggle(hook);
-    expect(hook.result.current[0]).toBeFalsy();
-    act(() => {
-      hook.result.current[1].setLeft();
-    });
-    expect(hook.result.current[0]).toBe('Hello');
-    act(() => {
-      hook.result.current[1].setRight();
-    });
-    expect(hook.result.current[0]).toBeFalsy();
-  });
-
-  it('test on optional', () => {
-    const hook = renderHook(() => useToggle('Hello', 'World'));
-    callToggle(hook);
-    expect(hook.result.current[0]).toBe('World');
-    act(() => {
-      hook.result.current[1].set('World');
-    });
-    expect(hook.result.current[0]).toBe('World');
-    callToggle(hook);
-    expect(hook.result.current[0]).toBe('Hello');
-  });
-});
