+++ packages/hooks/src/useToggle/tests/index.spec.ts
@@ -1,0 +1,43 @@
+import { act, renderHook } from '@testing-library/react';
+import { describe, expect, test } from 'vitest';
+import useToggle from '../index';
+
+const callToggle = (hook: any) => {
+  act(() => {
+    hook.result.current[1].toggle();
+  });
+};
+
+describe('useToggle', () => {
+  test('test on init', async () => {
+    const hook = renderHook(() => useToggle());
+    expect(hook.result.current[0]).toBeFalsy();
+  });
+
+  test('test on methods', async () => {
+    const hook = renderHook(() => useToggle('Hello'));
+    expect(hook.result.current[0]).toBe('Hello');
+    callToggle(hook);
+    expect(hook.result.current[0]).toBeFalsy();
+    act(() => {
+      hook.result.current[1].setLeft();
+    });
+    expect(hook.result.current[0]).toBe('Hello');
+    act(() => {
+      hook.result.current[1].setRight();
+    });
+    expect(hook.result.current[0]).toBeFalsy();
+  });
+
+  test('test on optional', () => {
+    const hook = renderHook(() => useToggle('Hello', 'World'));
+    callToggle(hook);
+    expect(hook.result.current[0]).toBe('World');
+    act(() => {
+      hook.result.current[1].set('World');
+    });
+    expect(hook.result.current[0]).toBe('World');
+    callToggle(hook);
+    expect(hook.result.current[0]).toBe('Hello');
+  });
+});
