--- packages/hooks/src/useUnmountedRef/__tests__/index.test.ts
@@ -1,13 +1,0 @@
-import { renderHook } from '../../utils/tests';
-import useUnmountedRef from '../index';
-
-describe('useUnmountedRef', () => {
-  it('should work', async () => {
-    const hook = renderHook(() => useUnmountedRef());
-    expect(hook.result.current.current).toBe(false);
-    hook.rerender();
-    expect(hook.result.current.current).toBe(false);
-    hook.unmount();
-    expect(hook.result.current.current).toBe(true);
-  });
-});
