--- packages/hooks/src/useWhyDidYouUpdate/__tests__/index.test.ts
@@ -1,33 +1,0 @@
-import { renderHook, act } from '@testing-library/react';
-import useWhyDidYouUpdate from '../index';
-import { useState } from 'react';
-
-describe('useWhyDidYouUpdate', () => {
-  it('should work', () => {
-    console.log = jest.fn();
-    const setup = () =>
-      renderHook(() => {
-        const [count, setCount] = useState(100);
-        useWhyDidYouUpdate('UseWhyDidYouUpdateComponent', { count });
-        return {
-          setCount,
-        };
-      });
-
-    const hook = setup();
-
-    act(() => {
-      hook.result.current.setCount(1);
-    });
-    expect(console.log).toHaveBeenCalledWith(
-      '[why-did-you-update]',
-      'UseWhyDidYouUpdateComponent',
-      {
-        count: {
-          from: 100,
-          to: 1,
-        },
-      },
-    );
-  });
-});
